+++ data/email-sign-in/src/main/java/com/apeun/gidaechi/data/MemberRepository.kt
@@ -5,6 +5,6 @@ import com.apeun.gidaechi.data.model.EmailSignInModel
 import com.apeun.gidaechi.network.request.EmailSignInRequest
 import kotlinx.coroutines.flow.Flow
 
-interface EmailSignInRepository {
+interface MemberRepository {
     suspend fun emailSignIn(body: EmailSignInRequest): Flow<Result<EmailSignInModel>>
 }
