--- network/sign-in/src/main/java/com/apeun/gidaechi/network/EmailSignUpDatasource.kt
@@ -1,8 +1,0 @@
-package com.apeun.gidaechi.network
-
-import com.apeun.gidaechi.network.core.response.Response
-import com.apeun.gidaechi.network.request.EmailSignUpReqest
-
-interface EmailSignUpDatasource {
-    suspend fun emailSignUp(body: EmailSignUpReqest): Response
-}
