+++ network/sign-in/src/main/java/com/apeun/gidaechi/network/MemberDatasource.kt
@@ -1,9 +1,12 @@
 package com.apeun.gidaechi.network
 
 import com.apeun.gidaechi.network.core.response.BaseResponse
+import com.apeun.gidaechi.network.core.response.Response
 import com.apeun.gidaechi.network.request.EmailSignInRequest
+import com.apeun.gidaechi.network.request.EmailSignUpReqest
 import com.apeun.gidaechi.network.response.EmailSignInResponse
 
-interface EmailSignInDatasource {
+interface MemberDatasource {
     suspend fun emailSignIn(body: EmailSignInRequest): BaseResponse<EmailSignInResponse>
+    suspend fun emailSignUp(body: EmailSignUpReqest): Response
 }
