+++ network/sign-in/src/main/java/com/apeun/gidaechi/network/datasource/MemberDatasourceImpl.kt
@@ -2,10 +2,12 @@ package com.apeun.gidaechi.network.datasource
 
 import com.apeun.gidaechi.common.utiles.DispatcherType
 import com.apeun.gidaechi.common.utiles.SeugiDispatcher
-import com.apeun.gidaechi.network.EmailSignInDatasource
+import com.apeun.gidaechi.network.MemberDatasource
 import com.apeun.gidaechi.network.core.SeugiUrl
 import com.apeun.gidaechi.network.core.response.BaseResponse
+import com.apeun.gidaechi.network.core.response.Response
 import com.apeun.gidaechi.network.request.EmailSignInRequest
+import com.apeun.gidaechi.network.request.EmailSignUpReqest
 import com.apeun.gidaechi.network.response.EmailSignInResponse
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
@@ -14,11 +16,16 @@ import io.ktor.client.request.setBody
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineDispatcher
 
-class EmailSignInDatasourceImpl @Inject constructor(
+class MemberDatasourceImpl @Inject constructor(
     @SeugiDispatcher(DispatcherType.IO) private val dispatcher: CoroutineDispatcher,
     private val httpClient: HttpClient,
-) : EmailSignInDatasource {
+) : MemberDatasource {
     override suspend fun emailSignIn(body: EmailSignInRequest): BaseResponse<EmailSignInResponse> = httpClient.post("${SeugiUrl.Member.EMAIL_SIGN_IN}") {
         setBody(body = body)
     }.body<BaseResponse<EmailSignInResponse>>()
+
+    override suspend fun emailSignUp(body: EmailSignUpReqest): Response =
+        httpClient.post("${SeugiUrl.Member.EMAIL_SIGN_UP}"){
+            setBody(body = body)
+        }.body()
 }
