+++ network/sign-in/src/main/java/com/apeun/gidaechi/network/di/MemberDatasourceModule.kt
@@ -1,9 +1,7 @@
 package com.apeun.gidaechi.network.di
 
-import com.apeun.gidaechi.network.EmailSignInDatasource
-import com.apeun.gidaechi.network.EmailSignUpDatasource
-import com.apeun.gidaechi.network.datasource.EmailSignInDatasourceImpl
-import com.apeun.gidaechi.network.datasource.EmailSignUpDatasourceImpl
+import com.apeun.gidaechi.network.MemberDatasource
+import com.apeun.gidaechi.network.datasource.MemberDatasourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -16,7 +14,7 @@ interface MemberDatasourceModule {
 
     @Binds
     @Singleton
-    fun provideEmailSignInDataSource(emailSignInDatasourceImpl: EmailSignInDatasourceImpl): EmailSignInDatasource
+    fun provideEmailSignInDataSource(emailSignInDatasourceImpl: MemberDatasourceImpl): MemberDatasource
 
     @Binds
     @Singleton
